requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML/CV
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0

# Redis for caching/queues
redis==5.0.1

# Async
asyncio==3.4.3

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1